reorganize sumcheck
